Begin refactoring repo: 1. Create fitting_helpers sub-package to fit_smah_helpers.py and related code 2. Do we even need stars.py? 3. Created defaults.py to store parameter dictionaries
