+++ diffstar/constants.py
@@ -4,3 +4,9 @@ import numpy as np
 
 TODAY = 13.8
 LGT0 = np.log10(TODAY)
+
+
+# Constants related to SFH integrals
+SFR_MIN = 1e-14
+T_BIRTH_MIN = 0.001
+N_T_LGSM_INTEGRATION = 100
