+++ diffstar/fitting_helpers/tests/test_stars.py
